requirements.txt
django-environ
gunicorn
pillow
mysqlclient>=2.2
django-ckeditor
babel